Move file and imported component into sign in page
